+++ packages/core/archive/components/Footer/Footer.data.js
