+++ packages/core/archive/components/Footer/FooterWithCols.data.js
