--- packages/core/stories/components/ErrorPage/ErrorPage.stories.mdx
@@ -1,30 +1,0 @@
-import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
-
-import generateTitle from '../../util/generateTitle';
-import * as stories from './ErrorPage.stories.js';
-
-<Meta
-  title={generateTitle('ErrorPage')}
-  parameters={{
-    viewMode: 'docs'
-  }}
-/>
-
-# Error Pages
-
-These are page templates designed to be displayed when an HTTP request is not successful and returns an error code. The most common codes that you will encounter are 404 (Not Found), 403 (Forbidden) and 500 (Internal Server Error).
-
-### 404 — Not Found
-<Canvas>
-  <Story story={stories.errorPage404} />
-</Canvas>
-
-### 403 — Forbidden
-<Canvas>
-  <Story story={stories.errorPage403} />
-</Canvas>
-
-### 500 — Internal Server Error
-<Canvas>
-  <Story story={stories.errorPage500} />
-</Canvas>
