--- packages/core/stories/hidden/components/Footer/Footer.data.js
@@ -1,12 +1,0 @@
-export default {
-  links: [{
-    href: 'https://www.mass.gov/topics/massachusetts-topics',
-    text: 'All Topics'
-  }, {
-    href: 'https://www.mass.gov/massgov-site-policies',
-    text: 'Site Policies'
-  }, {
-    href: 'https://www.mass.gov/topics/public-records-requests',
-    text: 'Public Records Requests'
-  }]
-};
