--- packages/core/stories/hidden/components/Footer/FooterWithCols.data.js
@@ -1,44 +1,0 @@
-export default {
-  items: [
-    {
-      heading: 'Footer Primary Links',
-      id: 'FooterLinks1',
-      links: [
-        {
-          href: 'https://www.mass.gov/topics/living',
-          text: 'Living'
-        },
-        {
-          href: 'https://www.mass.gov/topics/working',
-          text: 'Working'
-        },
-        {
-          href: 'https://www.mass.gov/topics/learning',
-          text: 'Learning'
-        },
-        {
-          href: 'https://www.mass.gov/topics/exploring-massachusetts',
-          text: 'Visiting & Exploring'
-        },
-        {
-          href: 'https://www.mass.gov/topics/your-government',
-          text: 'Your Government'
-        }
-      ]
-    },
-    {
-      heading: 'Footer Secondary Links',
-      id: 'FooterLinks2',
-      links: [
-        {
-          href: 'https://www.mass.gov/site-policies',
-          text: 'Site Policies'
-        },
-        {
-          href: 'https://www.mass.gov/topics/public-records-requests',
-          text: 'Public Records Requests'
-        }
-      ]
-    }
-  ]
-};
