--- packages/core/stories/hidden/elements/Link/Link.stories.mdx
@@ -1,101 +1,0 @@
-import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
-import Button from '@massds/mayflower-react/dist/Button';
-import DecorativeLink from '@massds/mayflower-react/dist/DecorativeLink';
-import CalloutLink from '@massds/mayflower-react/dist/CalloutLink';
-import generateTitle from '../../util/generateTitle';
-import { renderReactMarkup } from '../../util/renderCode';
-import * as stories from './Link.stories.js';
-
-<Meta
-  title={generateTitle('Link')}
-  parameters={{
-    viewMode: 'docs'
-  }}
-/>
-
-# Links
-
-Links are used to embed actions or pathways to more information in a sentence or on a page.
-
----
-
-### Inline Link
-
-- **Usage:** Use when you are linking just a few words of text or when you are doing inline links.
-- **Style:** Inline links take the style of the surrounding text, adding only the blue color and an underline.
-
-<Canvas>
-  <Story
-    name="link">
-      <p>Some content and then an <a href="#">inline link</a>, followed by more content.</p>
-  </Story>
-  <Story story={stories.linkCSS} />
-</Canvas>
-
----
-
-### Decorative Link
-
-- **Usage:** Used for standalone links that require more emphasis, usually inside `ma__section-links`.
-- **Style:** the decorative link is slightly larger in size and appended with an arrow icon.
-
-<Canvas>
-  <Story
-    name="link-decorative"
-    parameters={{ docs: { transformSource: () => renderReactMarkup(
-    <DecorativeLink text="Decorative Link"/>) } }}>
-    <DecorativeLink text="Decorative Link"/>
-  </Story>
-  <Story story={stories.linkDecorativeCSS} />
-</Canvas>
-
----
-
-### Callout Link
-
-- **Usage:** Call to action links that require more emphasis.
-- **Style:** The callout link has the style of a decorative link, but with a slightly larger font-size, and enlosed in a box.
-
-<Canvas>
-  <Story
-    name="link-callout"
-    parameters={{ docs: { transformSource: () => renderReactMarkup(<CalloutLink text="Callout Link"/>) } }}>
-      <CalloutLink text="Callout Link"/>
-  </Story>
-  <Story story={stories.linkCalloutCSS} />
-</Canvas>
-
----
-
-### Link Styled as a Button
-
-- **Usage:** If absolutely necessary this can be used for calls-to-action if there are good design reasons not to use Callout Link. In general, links should be styled differently than buttons, to help set expectations of their different behaviors.
-- **Style:** Styled exactly the same as `ma__button`.
-
-<Canvas>
-  <Story
-    name="link-button-primary"
-    parameters={{ docs: { transformSource: () => renderReactMarkup(<Button href="#" text="Link as a button" usage="primary" />) } }}>
-    	<Button href="#" text="Link as a button" usage="primary" />
-  </Story>
-  <Story
-    name="link-button-secondary"
-    parameters={{ docs: { transformSource: () => renderReactMarkup(<Button href="#" text="Link as a button" usage="secondary" />) } }}>
-    	<Button href="#" text="Link as a button" usage="secondary" />
-  </Story>
-  <Story
-    name="link-button-tertiary"
-    parameters={{ docs: { transformSource: () => renderReactMarkup(<Button href="#" text="Link as a button" usage="tertiary" />) } }}>
-    	<Button href="#" text="Link as a button" usage="tertiary" />
-  </Story>
-  <Story story={stories.linkButtonCSS} />
-</Canvas>
-
----
-
-### Accessibility & Best Practices
-
-The purpose of a link can be determined from the link text alone or from the text combined with programmatically determined content
-
-- Avoid using language in links that provides no specific information or is repetitive like “read more”.
-- Links that start a process like downloading files, opening your e-mail client, or calling a phone number should be clearly labeled for what they do.
