+++ packages/core/stories/tokens/colors/Colors.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs/blocks';
-import generateTitle from '../../util/generateTitle';
+import generateTitle from '@core/stories/util/generateTitle';
 import { ColorSwatch, GradientSpectrum } from './ColorDisplay';
 import { brandColors, grayScaleColors, utilityColors, primaryColors, primaryAltColors, highLightColors} from './colors.json';
