+++ packages/core/stories/tokens/logo/Logo.stories.mdx
@@ -4,7 +4,7 @@ import IconDownload from '@massds/mayflower-react/dist/Icon/IconDownload';
 import { ColorSwatch } from '../colors/ColorDisplay';
 import * as stories from './Logo.stories.js';
 
-import generateTitle from '../../util/generateTitle';
+import generateTitle from '@core/stories/util/generateTitle';
 const { STORYBOOK_CDN_PATH } = process.env;
 
 <Meta
